+++ bundles/framework/bundle/mapmodule-plugin/plugin/logo/LogoPlugin.js
@@ -19,9 +19,19 @@ Oskari.clazz.define('Oskari.mapframework.bundle.mapmodule.plugin.LogoPlugin',
     }, {
 
         templates: {
-            main: jQuery("<div class='logoplugin'><div class='icon'></div>" +
-                "<div class='terms'><a href='JavaScript:void(0);'></a></div>" +
-                "</div>")
+            main: jQuery(
+                "<div class='logoplugin'>" +
+                    "<div class='icon'></div>" +
+                    "<div class='terms'><a href='JavaScript:void(0);'></a></div>" +
+                    "<div class='data-sources'></div>" +
+                "</div>"
+            ),
+            dataSourcesDialog: jQuery(
+                "<div class='data-sources-dialog'>" +
+                    "<div class='layers'><h4></h4></div>" +
+                    "<div class='indicators'><h4></h4></div>" +
+                "</div>"
+            )
         },
 
         /** @static @property __name plugin name */
@@ -148,7 +158,11 @@ Oskari.clazz.define('Oskari.mapframework.bundle.mapmodule.plugin.LogoPlugin',
          * @property {Object} eventHandlers
          * @static
          */
-        eventHandlers: {},
+        eventHandlers: {
+            'StatsGrid.IndicatorsEvent': function(event) {
+                this._addIndicatorsToDataSourcesDialog(event.getIndicators());
+            }
+        },
 
         /** 
          * @method onEvent
@@ -156,7 +170,10 @@ Oskari.clazz.define('Oskari.mapframework.bundle.mapmodule.plugin.LogoPlugin',
          * Event is handled forwarded to correct #eventHandlers if found or discarded if not.
          */
         onEvent: function (event) {
-            return this.eventHandlers[event.getName()].apply(this, [event]);
+            var handler = this.eventHandlers[event.getName()];
+            if (handler) {
+                return handler.apply(this, [event]);
+            }
         },
 
         setLocation: function (location, logoContainer) {
@@ -202,7 +219,8 @@ Oskari.clazz.define('Oskari.mapframework.bundle.mapmodule.plugin.LogoPlugin',
                 link,
                 linkParams,
                 mapUrl,
-                termsUrl;
+                termsUrl,
+                dataSources;
 
             if (me.conf) {
                 mapUrl = sandbox.getLocalizedProperty(me.conf.mapUrlPrefix);
@@ -239,6 +257,17 @@ Oskari.clazz.define('Oskari.mapframework.bundle.mapmodule.plugin.LogoPlugin',
             } else {
                 link.hide();
             }
+
+            dataSources = me.element.find('div.data-sources');
+            if (me.conf && me.conf.hideDataSourceLink) {
+                dataSources.remove();
+            } else {
+                dataSources.html(myLoc.dataSources);
+                dataSources.click(function(e) {
+                    me._openDataSourcesDialog(e.target);
+                    me._requestDataSources();
+                });
+            }
         },
 
         /**
@@ -261,6 +290,114 @@ Oskari.clazz.define('Oskari.mapframework.bundle.mapmodule.plugin.LogoPlugin',
             testRegex = /oskari-publisher-font-/;
 
             this.getMapModule().changeCssClasses(classToAdd, testRegex, [div]);
+        },
+
+        /**
+         * Sends a request to get indicators. If the statsgrid bundle is not available
+         * (and consequently there aren't any indicators) it opens the data sources dialog
+         * and just shows the data sources of the layers.
+         *
+         * @method _requestDataSources
+         * @return {undefined}
+         */
+        _requestDataSources: function() {
+            var me = this,
+                reqBuilder = me._sandbox.getRequestBuilder('StatsGrid.IndicatorsRequest'),
+                request;
+
+            if (reqBuilder) {
+                request = reqBuilder();
+                me._sandbox.request(me, request);
+            }
+        },
+
+        /**
+         * Opens a dialog to show data sources of the selected layers
+         * and statistics indicators.
+         *
+         * @method _openDataSourcesDialog
+         * @param  {jQuery} target the target element where the popup is attached to
+         * @param  {Array[Object]} indicators the open indicators
+         * @return {undefined}
+         */
+        _openDataSourcesDialog: function(target) {
+            var me = this,
+                pluginLoc = me.getMapModule().getLocalization('plugin', true)[me.__name],
+                popupTitle = pluginLoc.dataSources,
+                dialog = Oskari.clazz.create('Oskari.userinterface.component.Popup'),
+                closeButton = Oskari.clazz.create('Oskari.userinterface.component.Button'),
+                content = me.templates.dataSourcesDialog.clone(),
+                layersCont = content.find('div.layers'),
+                layersHeaderLoc = pluginLoc.layersHeader,
+                layers = me._sandbox.findAllSelectedMapLayers(),
+                layersLen = layers.length,
+                layer, i;
+
+            closeButton.setTitle('OK');
+            closeButton.setHandler(function () {
+                me.dataSourcesDialog = null;
+                dialog.close(true);
+            });
+
+            // List the layers if any
+            if (layersLen === 0) {
+                layersCont.remove();
+            } else {
+                layersCont.find('h4').html(layersHeaderLoc);
+
+                for (i = 0; i < layersLen; ++i) {
+                    layer = layers[i];
+                    layersCont.append(
+                        '<div>' +
+                            layer.getName() + ' - ' + layer.getOrganizationName() +
+                        '</div>'
+                    );
+                }
+            }
+
+            this.dataSourcesDialog = dialog;
+
+            dialog.show(popupTitle, content, [closeButton]);
+
+            target = target || me.element.find('div.data-sources');
+            dialog.moveTo(target, 'top');
+        },
+
+        /**
+         * Adds indicators to the data sources dialog.
+         *
+         * @method _addIndicatorsToDataSourcesDialog
+         * @param {Object} indicators
+         */
+        _addIndicatorsToDataSourcesDialog: function(indicators) {
+            var dialog = this.dataSourcesDialog;
+            if (!dialog) return;
+
+            var pluginLoc = this.getMapModule().getLocalization('plugin', true)[this.__name],
+                content = dialog.getJqueryContent(),
+                indicatorsCont = content.find('div.indicators'),
+                indicatorsHeaderLoc = pluginLoc.indicatorsHeader,
+                indicators = indicators || {},
+                indicator, i, target;
+
+            // List the indicators if any
+            if (jQuery.isEmptyObject(indicators)) {
+                indicatorsCont.remove();
+            } else {
+                indicatorsCont.find('h4').html(indicatorsHeaderLoc);
+
+                for (i in indicators) {
+                    indicator = indicators[i];
+                    indicatorsCont.append(
+                        '<div>' +
+                            indicator.title + ' - ' + indicator.organization +
+                        '</div>'
+                    );
+                }
+            }
+
+            target = target || this.element.find('div.data-sources');
+            dialog.moveTo(target, 'top');
         }
     }, {
         /**